Add caching layer to reduce Strapi API calls
- Add request-level deduplication in api.ts to prevent duplicate calls
  within the same SSR request (e.g., site-settings called from Layout + page)
- Add Cache-Control headers to /api/posts and /api/climbing-ticks endpoints
- Create aggregated /api/ticklist-data endpoint that fetches all data
  server-side and pre-computes stats (reduces ~5-15 calls to 1)
- Create /api/subscribe proxy for newsletter signup
- Update ticklist-loader.ts to use new aggregated endpoint
- Update newsletter-signup.ts to use proxy endpoint

Closes #110
